oopsy sur strnequ mauvais ret

+++ includes/shared_libft.h
@@ -6,7 +6,7 @@
 /*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/07/07 12:36:29 by arsciand          #+#    #+#             */
-/*   Updated: 2020/02/11 15:11:54 by guvillat         ###   ########.fr       */
+/*   Updated: 2020/02/11 15:14:03 by guvillat         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -76,7 +76,7 @@ void				*ft_strlower(char *s);
 char				*ft_strncat(char *s1, const char *s2, size_t n);
 int					ft_strncmp(const char *s1, const char *s2, size_t n);
 char				*ft_strncpy(char *s1, const char *s2, size_t n);
-u_int8_t			ft_strnequ(const char *s1, const char *s2, size_t n);
+int				ft_strnequ(const char *s1, const char *s2, size_t n);
 char				*ft_strnew(size_t size);
 char				*ft_strrchr(const char *s, int c);
 char				**ft_strsplit(const char *s, char *charset);
